Split the code into multiple files
