--- accessibility-checker-for-vaadin/src/test/java/org/vaadin/addons/accessibility/AccessibilityJavaSourceModifierTest.java
@@ -1,7 +1,0 @@
-package org.vaadin.addons.accessibility;
-
-/**
- * @author jcgueriaud
- */
-public class AccessibilityJavaSourceModifierTest  extends AbstractClassBasedTest {
-}
